Major refactor to introduce a new way to handle missing last event id
In former implementation, resuming with an event id no longer in the capped
collection was leading to a resuming failure, forcing the consumer to either
restart a full replication or live with an inconsistent view of the dataset.

In order to prevent this from happening, we had to have a big enough capped
collection. Having millions of operation in this capped collection and a lot
of consumers easily led to a lock contention at MongoDb level.

This refactor introduce the fact that a missing event id in the capped
collection can now fallback to a resume in the replication at the timestamp
of requested event id. We use the ObjectId embeded timestamp to bridge the
two systems together. This have the added advantage of removing all the
intermediate operations when a consumer a lagging behind, as replication
does only send final state operations.

As this change implies a smaller capped collection and some change in the
format of the stored object, the agent use diffent collection names. This
makes this version non backward compatible in term of data storage. You
can use the oplog-sync to perform the migration though.